pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "odoodev-equitania"
version = "0.1.0"
description = "Unified CLI tool for native Odoo development environment management"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    { name = "Equitania Software GmbH", email = "info@equitania.de" },
]
keywords = ["odoo", "development", "cli", "native", "devtool"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Environment :: Console",
]
dependencies = [
    "click>=8.1.7",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "Jinja2>=3.1.0",
]

[project.scripts]
odoodev = "odoodev.cli:cli"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "build>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["odoodev"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["odoodev"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
